Design: Weather by city and location

+++ SunnyXamarin/Models/WeatherModel.cs
@@ -47,16 +47,16 @@ namespace SunnyXamarin.Models
             //{"temp":286.32,"feels_like":285.42,"temp_min":283.61,"temp_max":286.64,"pressure":1011,"humidity":66}
 
             [JsonProperty("temp")]
-            public float temp { get; set; }
+            public double temp { get; set; }
 
             [JsonProperty("feels_like")]
-            public float feels_like { get; set; }
+            public double feels_like { get; set; }
 
             [JsonProperty("temp_min")]
-            public float temp_min { get; set; }
+            public double temp_min { get; set; }
 
             [JsonProperty("temp_max")]
-            public float temp_max { get; set; }
+            public double temp_max { get; set; }
 
             [JsonProperty("pressure")]
             public int pressure { get; set; }
